backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.12.5
pydantic[email]==2.12.5
python-dotenv==1.0.1
python-multipart==0.0.12
supabase==2.24.0
email-validator==2.1.1
google-genai  # For Google Gemini models
yfinance>=0.2.0  # For stock price data
aiohttp>=3.9.0  # For async HTTP requests
PyJWT>=2.8.0  # For JWT token decoding
# Other LLM providers (optional)
# openai>=1.0.0  # For OpenAI GPT models
# anthropic>=0.18.0  # For Anthropic Claude models
# ollama>=0.1.0  # For local Ollama models

